backend/requirements.txt
requests
beautifulsoup4
python-dotenv
google-cloud-storage
minio
supabase
google-generativeai
openai
flask
flask-cors